Added method to see iterator level of an ExperimentIterator.
Example structure:

it_3
-->it_2
---->it_1
------>exp_1,exp_2,...

it_3 has level of 3 since there are two iterators before it and then one layer of normal experiments. it_2 has a level of 2. In genral the level is equal to the number of layers below it. This should help track the complicated onion someone might make.

To accomplish the level use the method detect_iterator_depth in init. It takes the experiments of an iteratro and checks if any of them has the method get_iterator_type which is exclusive to ExperimentIterators ie normal experiments dont have it. If there is an iterator it increases the level counter by 1 and goes into the sub iterator experiment to again check it is has a sub sub iterator.

May cause error for very long iterator trains but should ensure that EVERY iterator level knows what depth it is at. Reason for implementing is to see how far data will be nested in self.data dictionary. May have more uses in the future.